JP2Grok: refactor generic code into base driver
