+++ frmts/drivers.ini
@@ -72,6 +72,7 @@ JPIPKAK
 JP2Lura
 ECW
 JP2ECW
+JP2Grok
 JP2OpenJPEG
 L1B
 FIT
