+++ frmts/grok/makefile.vc
@@ -1,5 +1,5 @@
 
-OBJ	=	openjpegdataset.obj
+OBJ	=	grokdataset.obj jp2dataset.obj
 
 GDAL_ROOT	=	..\..
